requirements.txt
boto
boto3
botocore
numpy
pillow